+++ src/main/java/ru/vldf/sportsportal/dao/generic/definite/tourney/TeamCompositionDAO.java
@@ -1,9 +1,9 @@
 package ru.vldf.sportsportal.dao.generic.definite.tourney;
 
-import ru.vldf.sportsportal.model.tourney.TeamCompositionEntity;
-import ru.vldf.sportsportal.model.tourney.TeamCompositionStatusEntity;
-import ru.vldf.sportsportal.model.tourney.TeamEntity;
-import ru.vldf.sportsportal.model.tourney.TourneyEntity;
+import ru.vldf.sportsportal.domain.tourney.TeamCompositionEntity;
+import ru.vldf.sportsportal.domain.tourney.TeamCompositionStatusEntity;
+import ru.vldf.sportsportal.domain.tourney.TeamEntity;
+import ru.vldf.sportsportal.domain.tourney.TourneyEntity;
 
 import java.util.List;
