+++ src/main/java/ru/vldf/sportsportal/dao/generic/definite/tourney/TeamStatusDAO.java
@@ -1,6 +1,6 @@
 package ru.vldf.sportsportal.dao.generic.definite.tourney;
 
-import ru.vldf.sportsportal.model.tourney.TeamStatusEntity;
+import ru.vldf.sportsportal.domain.tourney.TeamStatusEntity;
 
 import java.util.List;
