+++ src/main/java/ru/vldf/sportsportal/dao/impl/lease/SportDAOImpl.java
@@ -3,7 +3,7 @@ package ru.vldf.sportsportal.dao.impl.lease;
 import org.springframework.stereotype.Repository;
 import ru.vldf.sportsportal.dao.generic.abstrct.AbstractDAOImpl;
 import ru.vldf.sportsportal.dao.generic.definite.lease.SportDAO;
-import ru.vldf.sportsportal.model.lease.SportEntity;
+import ru.vldf.sportsportal.domain.common.SportEntity;
 
 import java.util.List;
