+++ src/main/java/ru/vldf/sportsportal/dao/impl/tourney/TeamCompositionStatusDAOImpl.java
@@ -3,7 +3,7 @@ package ru.vldf.sportsportal.dao.impl.tourney;
 import org.springframework.stereotype.Repository;
 import ru.vldf.sportsportal.dao.generic.abstrct.AbstractDAOImpl;
 import ru.vldf.sportsportal.dao.generic.definite.tourney.TeamCompositionStatusDAO;
-import ru.vldf.sportsportal.model.tourney.TeamCompositionStatusEntity;
+import ru.vldf.sportsportal.domain.tourney.TeamCompositionStatusEntity;
 
 import java.util.List;
