+++ src/main/java/ru/vldf/sportsportal/dao/impl/tourney/TeamPlayerDAOImpl.java
@@ -3,8 +3,7 @@ package ru.vldf.sportsportal.dao.impl.tourney;
 import org.springframework.stereotype.Repository;
 import ru.vldf.sportsportal.dao.generic.abstrct.AbstractDAOImpl;
 import ru.vldf.sportsportal.dao.generic.definite.tourney.TeamPlayerDAO;
-import ru.vldf.sportsportal.model.tourney.TeamCompositionEntity;
-import ru.vldf.sportsportal.model.tourney.TeamPlayerEntity;
+import ru.vldf.sportsportal.domain.tourney.TeamPlayerEntity;
 
 import java.util.List;
