+++ src/main/java/ru/vldf/sportsportal/dao/impl/tourney/TeamStatusDAOImpl.java
@@ -3,7 +3,7 @@ package ru.vldf.sportsportal.dao.impl.tourney;
 import org.springframework.stereotype.Repository;
 import ru.vldf.sportsportal.dao.generic.abstrct.AbstractDAOImpl;
 import ru.vldf.sportsportal.dao.generic.definite.tourney.TeamStatusDAO;
-import ru.vldf.sportsportal.model.tourney.TeamStatusEntity;
+import ru.vldf.sportsportal.domain.tourney.TeamStatusEntity;
 
 import java.util.List;
