+++ src/main/java/ru/vldf/sportsportal/dao/impl/user/UserRoleDAOImpl.java
@@ -3,7 +3,7 @@ package ru.vldf.sportsportal.dao.impl.user;
 import org.springframework.stereotype.Repository;
 import ru.vldf.sportsportal.dao.generic.definite.user.UserRoleDAO;
 import ru.vldf.sportsportal.dao.generic.abstrct.AbstractDAOImpl;
-import ru.vldf.sportsportal.model.user.UserRoleEntity;
+import ru.vldf.sportsportal.domain.user.UserRoleEntity;
 
 import java.util.List;
