+++ src/main/java/ru/vldf/sportsportal/domain/tourney/TeamCompositionEntity.java
@@ -1,4 +1,4 @@
-package ru.vldf.sportsportal.model.tourney;
+package ru.vldf.sportsportal.domain.tourney;
 
 import ru.vldf.sportsportal.dto.tourney.TeamCompositionDTO;
