+++ src/main/java/ru/vldf/sportsportal/domain/tourney/TeamPlayerEntity.java
@@ -1,7 +1,7 @@
-package ru.vldf.sportsportal.model.tourney;
+package ru.vldf.sportsportal.domain.tourney;
 
 import ru.vldf.sportsportal.dto.tourney.TeamPlayerDTO;
-import ru.vldf.sportsportal.model.user.UserEntity;
+import ru.vldf.sportsportal.domain.user.UserEntity;
 
 import javax.persistence.*;
 import java.sql.Date;
