+++ src/main/java/ru/vldf/sportsportal/dto/common/SportDTO.java
@@ -1,0 +1,66 @@
+package ru.vldf.sportsportal.dto.common;
+
+import ru.vldf.sportsportal.domain.common.SportEntity;
+
+public class SportDTO {
+    private Integer id;
+    private String code;
+    private String description;
+
+    public SportDTO() {
+
+    }
+
+    public SportDTO(SportEntity sport) {
+        id = sport.getId();
+        code = sport.getCode();
+        description = sport.getDescription();
+    }
+
+    public Integer getId() {
+        return id;
+    }
+
+    public void setId(Integer id) {
+        this.id = id;
+    }
+
+    public String getCode() {
+        return code;
+    }
+
+    public void setCode(String name) {
+        this.code = name;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+
+        SportDTO sportDTO = (SportDTO) o;
+
+        return id != null ? id.equals(sportDTO.id) : sportDTO.id == null;
+    }
+
+    @Override
+    public int hashCode() {
+        return id != null ? id.hashCode() : 0;
+    }
+
+    @Override
+    public String toString() {
+        return "SportDTO{" +
+                "id=" + id +
+                ", code='" + code + '\'' +
+                '}';
+    }
+}
