+++ src/main/java/ru/vldf/sportsportal/dto/lease/PlaygroundDTO.java
@@ -1,6 +1,6 @@
 package ru.vldf.sportsportal.dto.lease;
 
-import ru.vldf.sportsportal.model.lease.PlaygroundEntity;
+import ru.vldf.sportsportal.domain.lease.PlaygroundEntity;
 
 public class PlaygroundDTO {
     private Integer id;
@@ -39,5 +39,28 @@ public class PlaygroundDTO {
 
     public String getAddress() {
         return address;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+
+        PlaygroundDTO that = (PlaygroundDTO) o;
+
+        return id != null ? id.equals(that.id) : that.id == null;
+    }
+
+    @Override
+    public int hashCode() {
+        return id != null ? id.hashCode() : 0;
+    }
+
+    @Override
+    public String toString() {
+        return "PlaygroundDTO{" +
+                "id=" + id +
+                ", name='" + name + '\'' +
+                '}';
     }
 }
