+++ src/main/java/ru/vldf/sportsportal/dto/tourney/TeamDTO.java
@@ -1,7 +1,7 @@
 package ru.vldf.sportsportal.dto.tourney;
 
 import ru.vldf.sportsportal.dto.user.UserDTO;
-import ru.vldf.sportsportal.model.tourney.TeamEntity;
+import ru.vldf.sportsportal.domain.tourney.TeamEntity;
 
 public class TeamDTO {
     private Integer id;
