+++ src/main/java/ru/vldf/sportsportal/dto/tourney/TourneyStatusDTO.java
@@ -1,6 +1,6 @@
 package ru.vldf.sportsportal.dto.tourney;
 
-import ru.vldf.sportsportal.model.tourney.TourneyStatusEntity;
+import ru.vldf.sportsportal.domain.tourney.TourneyStatusEntity;
 
 public class TourneyStatusDTO {
     private Integer id;
