+++ src/main/java/ru/vldf/sportsportal/service/AuthService.java
@@ -9,8 +9,7 @@ import org.springframework.transaction.annotation.Transactional;
 import ru.vldf.sportsportal.dao.generic.definite.user.UserRoleDAO;
 import ru.vldf.sportsportal.dao.generic.definite.user.UserDAO;
 import ru.vldf.sportsportal.dto.user.UserDTO;
-import ru.vldf.sportsportal.model.user.UserRoleEntity;
-import ru.vldf.sportsportal.model.user.UserEntity;
+import ru.vldf.sportsportal.domain.user.UserEntity;
 import ru.vldf.sportsportal.service.security.SecurityPrincipal;
 
 @Service
