+++ src/main/java/ru/vldf/sportsportal/service/LeaseService.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.vldf.sportsportal.dao.generic.definite.lease.PlaygroundDAO;
 import ru.vldf.sportsportal.dto.lease.PlaygroundDTO;
-import ru.vldf.sportsportal.model.lease.PlaygroundEntity;
+import ru.vldf.sportsportal.domain.lease.PlaygroundEntity;
 
 import java.util.ArrayList;
 import java.util.List;
