+++ src/main/java/ru/vldf/sportsportal/service/UserService.java
@@ -9,8 +9,8 @@ import ru.vldf.sportsportal.dto.tourney.TeamCompositionDTO;
 import ru.vldf.sportsportal.dto.tourney.TeamDTO;
 import ru.vldf.sportsportal.dto.tourney.TeamPlayerDTO;
 import ru.vldf.sportsportal.dto.user.UserDTO;
-import ru.vldf.sportsportal.model.tourney.*;
-import ru.vldf.sportsportal.model.user.UserEntity;
+import ru.vldf.sportsportal.domain.tourney.*;
+import ru.vldf.sportsportal.domain.user.UserEntity;
 
 import java.util.ArrayList;
 import java.util.List;
