+++ src/main/java/ru/vldf/sportsportal/service/security/SecurityPrincipal.java
@@ -3,7 +3,7 @@ package ru.vldf.sportsportal.service.security;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.User;
 import ru.vldf.sportsportal.dto.user.UserDTO;
-import ru.vldf.sportsportal.model.user.UserEntity;
+import ru.vldf.sportsportal.domain.user.UserEntity;
 
 import java.util.Collection;
